+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 16.2.0 - 2026-03-25
+
 ### Added
 - Added a `--skipTypecheck` CLI option to skip typechecking before tests.